lab1src/requirements.txt
requests
streamlit
openai
cohere
langdetect
python-dotenv